+++ src/components/layouts/TwoColumnLayout/index.tsx
@@ -4,10 +4,10 @@ import { AsideContainer } from "@/components/atoms/containers/AsideContainer";
 import { BottomAreaContainer } from "@/components/atoms/containers/BottomAreaContainer";
 import { ContainerWithOrder } from "@/components/atoms/containers/ContainerWithOrder";
 import { MiddleAreaContainer } from "@/components/atoms/containers/MiddleAreaContainer";
-import { ShareButtons } from "@/components/atoms/ShareButtons";
 import { HtmlHeadBase } from "@/components/functions/meta";
 import { RootLayout } from "@/components/layouts/RootLayout";
 import BackLinks from "@/components/molecules/BackLinks";
+import { ShareButtons } from "@/components/molecules/ShareButtons";
 import { PickupArticles, PopularArticles } from "@/components/organisms/article";
 import { useMedia } from "@/hooks/useMedia/index";
 import type { TCategory, TConfig, TPickup, TRankedArticle } from "@/types";
