+++ src/components/molecules/ShareButtons/index.stories.tsx
@@ -3,7 +3,7 @@ import type { ComponentMeta, ComponentStory } from "@storybook/react";
 import { ShareButtons } from ".";
 
 export default {
-  title: "Atoms/ShareButtons",
+  title: "molecules/ShareButtons",
   component: ShareButtons,
   argTypes: {
     direction: { control: "radio", options: ["row", "column"] },
