+++ src/components/molecules/article/ArticleSkeltonList/ArticleSkelton/index.stories.tsx
@@ -3,7 +3,7 @@ import type { ComponentMeta, ComponentStory } from "@storybook/react";
 import { ArticleSkelton } from ".";
 
 export default {
-  title: "molecules/article/ArticleSkelton",
+  title: "molecules/article/ArticleSkeltonList/ArticleSkelton",
   component: ArticleSkelton,
 } as ComponentMeta<typeof ArticleSkelton>;
