+++ src/components/organisms/Header/index.tsx
@@ -1,9 +1,9 @@
 import { memo } from "react";
 
 import { SiteTitle } from "@/components/atoms/SiteTitle";
-import { ThemeSwitch } from "@/components/atoms/ThemeSwitch";
 import { CategoryMenu } from "@/components/molecules/category/CategoryMenu";
 import SearchBar from "@/components/molecules/SearchBar";
+import { ThemeSwitch } from "@/components/molecules/ThemeSwitch";
 import type { TCategory } from "@/types";
 
 type Props = {
